move follows and likes db access to separate clasess
