+++ backend/storage/user.go
@@ -16,15 +16,17 @@ import (
 
 // Struct that implements UserDataAcessor using given DAO and cache
 type UserStorage struct {
-	DAO   database.UserDAO
-	cache cache.CacheProvider
+	userDAO    database.UserDAO
+	followsDAO database.FollowsDAO
+	cache      cache.CacheProvider
 }
 
 // Constructs UserDB that uses a DAO and CacheProvider
-func NewUserStorage(DAO database.UserDAO, cache cache.CacheProvider) *UserStorage {
+func NewUserStorage(userDAO database.UserDAO, followsDAO database.FollowsDAO, cache cache.CacheProvider) *UserStorage {
 	return &UserStorage{
-		DAO,
-		cache,
+		userDAO:    userDAO,
+		followsDAO: followsDAO,
+		cache:      cache,
 	}
 }
 
@@ -34,18 +36,18 @@ func (s *UserStorage) GetUsers(requestingUserID int64) ([]*model.PublicUser, err
 		return users, nil
 	}
 
-	users, err := s.DAO.GetPublicUsers()
+	users, err := s.userDAO.GetPublicUsers()
 	if err != nil {
 		return nil, errors.UnexpectedError
 	}
 
 	// TODO: this could be done in parallel by a few goroutines...
 	for _, user := range users {
-		followerCount, err := s.DAO.FollowerCount(user.ID)
+		followerCount, err := s.followsDAO.FollowerCount(user.ID)
 		if err != nil {
 			return nil, errors.UnexpectedError
 		}
-		following, err := s.DAO.IsFollowing(requestingUserID, user.ID)
+		following, err := s.followsDAO.IsFollowing(requestingUserID, user.ID)
 		if err != nil {
 			return nil, errors.UnexpectedError
 		}
@@ -64,7 +66,7 @@ func (s *UserStorage) GetUserByID(userID, requestingUserID int64) (*model.Public
 		return user, nil
 	}
 
-	user, err := s.DAO.GetPublicUserWithID(userID)
+	user, err := s.userDAO.GetPublicUserWithID(userID)
 
 	if err == sql.ErrNoRows {
 		return nil, errors.NoResultsError
@@ -74,11 +76,11 @@ func (s *UserStorage) GetUserByID(userID, requestingUserID int64) (*model.Public
 		return nil, errors.UnexpectedError
 	}
 
-	followerCount, err := s.DAO.FollowerCount(user.ID)
+	followerCount, err := s.followsDAO.FollowerCount(user.ID)
 	if err != nil {
 		return nil, errors.UnexpectedError
 	}
-	following, err := s.DAO.IsFollowing(requestingUserID, user.ID)
+	following, err := s.followsDAO.IsFollowing(requestingUserID, user.ID)
 	if err != nil {
 		return nil, errors.UnexpectedError
 	}
@@ -96,7 +98,7 @@ func (s *UserStorage) GetUserByEmail(email string) (*model.User, error) {
 		return user, nil
 	}
 
-	user, err := s.DAO.GetUserWithEmail(email)
+	user, err := s.userDAO.GetUserWithEmail(email)
 	if err == sql.ErrNoRows {
 		return nil, errors.NoResultsError
 	}
@@ -109,7 +111,7 @@ func (s *UserStorage) GetUserByEmail(email string) (*model.User, error) {
 }
 
 func (s *UserStorage) InsertUser(newUserForm *model.NewUserForm) (*model.PublicUser, error) {
-	insertedUser, err := s.DAO.InsertUser(newUserForm)
+	insertedUser, err := s.userDAO.InsertUser(newUserForm)
 
 	if err != nil {
 		if err, ok := err.(*pq.Error); ok && err.Code == database.UniqueConstraintViolationCode {
@@ -126,7 +128,7 @@ func (s *UserStorage) InsertUser(newUserForm *model.NewUserForm) (*model.PublicU
 }
 
 func (s *UserStorage) UpdateUserLastLoginTime(userID int64, lastLoginTime *time.Time) error {
-	err := s.DAO.UpdateUserLastLoginTime(userID, lastLoginTime)
+	err := s.userDAO.UpdateUserLastLoginTime(userID, lastLoginTime)
 	if err != nil {
 		return errors.UnexpectedError
 	}
@@ -138,7 +140,7 @@ func (s *UserStorage) UpdateUserLastLoginTime(userID int64, lastLoginTime *time.
 }
 
 func (s *UserStorage) FollowUser(followeeID, followerID int64) error {
-	err := s.DAO.FollowUser(followeeID, followerID)
+	err := s.followsDAO.FollowUser(followeeID, followerID)
 	if err != nil {
 		return errors.UnexpectedError
 	}
@@ -151,7 +153,7 @@ func (s *UserStorage) FollowUser(followeeID, followerID int64) error {
 }
 
 func (s *UserStorage) UnfollowUser(followeeID, followerID int64) error {
-	err := s.DAO.UnfollowUser(followeeID, followerID)
+	err := s.followsDAO.UnfollowUser(followeeID, followerID)
 	if err != nil {
 		return errors.UnexpectedError
 	}
@@ -164,7 +166,7 @@ func (s *UserStorage) UnfollowUser(followeeID, followerID int64) error {
 }
 
 func (s *UserStorage) Followers(userID, requestingUserID int64) ([]*model.PublicUser, error) {
-	followersIDs, err := s.DAO.IDsOfFollowers(userID)
+	followersIDs, err := s.followsDAO.IDsOfFollowers(userID)
 	if err != nil {
 		return nil, errors.UnexpectedError
 	}
@@ -183,7 +185,7 @@ func (s *UserStorage) Followers(userID, requestingUserID int64) ([]*model.Public
 	}
 
 	if len(followersIDs) > 0 {
-		dbFollowers, err := s.DAO.GetPublicUsersFromListOfIDs(followersIDs)
+		dbFollowers, err := s.userDAO.GetPublicUsersFromListOfIDs(followersIDs)
 		if err != nil {
 			return nil, errors.UnexpectedError
 		}
@@ -194,7 +196,7 @@ func (s *UserStorage) Followers(userID, requestingUserID int64) ([]*model.Public
 }
 
 func (s *UserStorage) Followees(userID, requestingUserID int64) ([]*model.PublicUser, error) {
-	followeesIDs, err := s.DAO.IDsOfFollowees(userID)
+	followeesIDs, err := s.followsDAO.IDsOfFollowees(userID)
 	if err != nil {
 		return nil, errors.UnexpectedError
 	}
@@ -213,17 +215,17 @@ func (s *UserStorage) Followees(userID, requestingUserID int64) ([]*model.Public
 	}
 
 	if len(followeesIDs) > 0 {
-		dbFollowees, err := s.DAO.GetPublicUsersFromListOfIDs(followeesIDs)
+		dbFollowees, err := s.userDAO.GetPublicUsersFromListOfIDs(followeesIDs)
 		if err != nil {
 			return nil, errors.UnexpectedError
 		}
 
 		for _, user := range dbFollowees {
-			followerCount, err := s.DAO.FollowerCount(user.ID)
+			followerCount, err := s.followsDAO.FollowerCount(user.ID)
 			if err != nil {
 				return nil, errors.UnexpectedError
 			}
-			following, err := s.DAO.IsFollowing(requestingUserID, user.ID)
+			following, err := s.followsDAO.IsFollowing(requestingUserID, user.ID)
 			if err != nil {
 				return nil, errors.UnexpectedError
 			}
